--- Plugins/iOS/JPushUnityManager.h
@@ -1,17 +1,0 @@
-//
-//  JPushUnityAnalytics.h
-//  test_certifacate
-//
-//  Created by qinghe on 14-4-15.
-//  Copyright (c) 2014年 jpush. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-@interface JPushUnityManager : NSObject
-
-@end
-//extern id APNativeJSONObject(NSData *data);
-//extern NSData *APNativeJSONData(id obj);
-
-
